+++ src/types/get-deploy/static.ts
@@ -30,3 +30,20 @@ export type MendatoryModuleType =
  * @description Module name by module type for the deployment
  */
 export type ModuleNameByType = GetModuleNameByType<ModuleType>
+
+/**
+ * @description The return type for the fetchDeployment function
+ */
+export type FetchDeploymentReturnType = {
+  bancorFormula: Record<string, `0x${string}` | undefined>
+  erc20Mock: Record<string, `0x${string}` | undefined>
+  orchestratorFactory: Record<string, `0x${string}` | undefined>
+  restrictedPimFactory: Record<string, `0x${string}` | undefined>
+  immutablePimFactory: Record<string, `0x${string}` | undefined>
+  migratingPimFactory: Record<string, `0x${string}` | undefined>
+}
+
+/**
+ * @description The version of the deployment
+ */
+export type DeploymentVersion = 'v1.0.0'
