+++ src/utils/tagProcessor/allowance.ts
@@ -1,4 +1,4 @@
-import type { TagProcessorAllowancesParams } from '@'
+import type { TagProcessorAllowancesParams } from '@/index'
 import { ERC20_ABI } from '../constants'
 
 export default async function allowance({
