+++ tests/e2e/get-deploy/pim/default.test.ts
@@ -1,6 +1,6 @@
 import { expect, describe, it } from 'bun:test'
 
-import { type GetUserArgs, type RequestedModules, type Workflow } from '@'
+import { type GetUserArgs, type RequestedModules, type Workflow } from '@/index'
 import { getModuleSchema } from '@/get-deploy/get-inputs'
 import { getContract, isAddress, isHash, parseUnits } from 'viem'
 import {
