+++ tests/e2e/workflow/non-typesafe.test.ts
@@ -3,7 +3,7 @@ import { expect, describe, it } from 'bun:test'
 import { sdk } from 'tests/helpers'
 import { TEST_ERC20_MOCK_ADDRESS } from '../../helpers/constants'
 
-import type { Workflow } from '@'
+import type { Workflow } from '@/index'
 
 describe('#WORKFLOW_NON_TYPESAFE', () => {
   let workflow: Workflow<typeof sdk.walletClient, undefined>
